reduxでfetchDotsのactionを作成 #73
