+++ src/components/templates/MiniDots.jsx
@@ -1,48 +1,36 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
+import { useSelector } from "react-redux";
 import { makeStyles } from "@material-ui/core/styles";
-import firebase from "../../firebase/firebase";
 import List from "@material-ui/core/List";
 import Dots from "./Dots";
 
 const useStyles = makeStyles((theme) => ({
-  root: {
-    width: "100%",
-    height: "50vh",
-    maxWidth: 400,
-    backgroundColor: theme.palette.background.paper,
-    margin: "auto",
-    marginTop: 10,
-    position: "relative",
-  },
-  primary: {
-    whiteSpace: "nowrap",
-    overflow: "hidden",
-    textOverflow: "ellipsis",
-  },
-  addButton: {
-    fontSize: 30,
-    position: "fixed",
-    right: "0%",
-    bottom: "7%",
-  },
+	root: {
+		width: "100%",
+		height: "50vh",
+		maxWidth: 400,
+		backgroundColor: theme.palette.background.paper,
+		margin: "auto",
+		marginTop: 10,
+		position: "relative",
+	},
+	primary: {
+		whiteSpace: "nowrap",
+		overflow: "hidden",
+		textOverflow: "ellipsis",
+	},
+	addButton: {
+		fontSize: 30,
+		position: "fixed",
+		right: "0%",
+		bottom: "7%",
+	},
 }));
 
 export default function MiniDots() {
-	const [dots, setDots] = useState([]);
+	const dots = useSelector((state) => state.dots);
+	console.log(dots);
 	const classes = useStyles();
-
-	useEffect(() => {
-		firebase
-			.firestore()
-			.collection("dots")
-			.onSnapshot((snapshot) => {
-				setDots(
-					snapshot.docs.map((doc) => {
-						return doc.data();
-					})
-				);
-			});
-	}, []);
 
 	return (
 		<>
