+++ src/reducks/dots/action.jsx
@@ -5,3 +5,12 @@ export const add_dot = (dot) => {
 		payload: dot,
 	};
 };
+
+export const FETCH_DOT = "FETCH_DOT";
+export const fetch_dot = (dot) => {
+	console.log(dot)
+	return {
+		type: FETCH_DOT,
+		payload: dot,
+	};
+};
