+++ src/reducks/dots/reducers.jsx
@@ -5,6 +5,8 @@ export const DotReducer = (state = initialState, action) => {
 	switch (action.type) {
 		case Actions.ADD_DOT:
 			return [...state, action.payload];
+		case Actions.FETCH_DOT:
+			return action.payload;
 		default:
 			return state;
 	}
